menu stuffs

+++ src/components/Editor/index.tsx
@@ -1,26 +1,26 @@
-import Block from "./Block"
 import { DetailedHTMLProps, HTMLAttributes, KeyboardEventHandler, MouseEventHandler, useEffect, useState } from "react"
 import { FC } from "react"
 import styles from "styles/components/Editor.module.scss"
-import useStatePromise from "utils/useStatePromise"
-import { DragDropContext, Draggable, Droppable, DropResult, ResponderProvided } from "react-beautiful-dnd"
-import { BlockInfo, getNewBlock } from "utils/editor/block"
+import { DragDropContext, Droppable, DropResult, ResponderProvided } from "react-beautiful-dnd"
+import { Block, createBlock } from "utils/editor/block"
+import BlockComponent from "./BlockComponent"
 
 type DivProps = DetailedHTMLProps<HTMLAttributes<HTMLDivElement>, HTMLDivElement>
 type Props = DivProps;
 
 const Editor: FC<Props> = ({ className, ...props }) => {
-  const [dragging, setDragging] = useStatePromise(false);
-  const [blocks, setBlocks] = useState<BlockInfo[]>([])
-  const [row, setRow] = useStatePromise(-1);
-  const [col, setCol] = useStatePromise<[number, number]>([-1, 0]);
-  const [selected, setSelected] = useStatePromise(-1);
+  const [dragging, setDragging] = useState(false);
+  const [blocks, setBlocks] = useState<Block[]>([
+    createBlock({ type: "title" })
+  ])
 
-  const dragStart = () => {
-    setDragging(true);
-  }
+  const [row, setRow] = useState(0);
+  const [col, setCol] = useState<[number, number]>([0, 0]);
+  const [selected, setSelected] = useState(-1);
+
+  const dragStart = () => setDragging(true);
   
-  const handleDrop = (result: DropResult, _provided: ResponderProvided) => {
+  const dragEnd = (result: DropResult, _provided: ResponderProvided) => {
     if (!result.destination) return;
     let updatedList = [...blocks];
     const [reorderedItem] = updatedList.splice(result.source.index, 1);
@@ -30,48 +30,47 @@ const Editor: FC<Props> = ({ className, ...props }) => {
     setBlocks(updatedList);
   };
 
-  const modifyBlock = (index: number, info: Partial<BlockInfo>) => {
+  const modifyBlock = (index: number, info: Partial<Block>) => {
     setBlocks(blocks => blocks.map((block, i) => {
-      return index == i ? {
-        ...block,
-        ...info
-      } : block
+      return index == i ? Object.assign(block, info) : block
     }))
   }
 
   const deleteCurrentBlock = () => {
     if(row > 0) {
-      const temp = blocks.map(block => ({...block}));
-
-      temp[row-1].content += temp[row].content;
-      temp.splice(row, 1)
-
       setRow(row-1)
-      setCol([blocks[row-1].content.length || 0, 0])
-      setBlocks(temp)
+      setCol([blocks[row-1].content.length, 0])
+
+      setBlocks(blocks => {
+        blocks[row-1].content += blocks[row].content;
+        blocks.splice(row, 1)
+        return blocks;
+      })
     }
   }
 
   const addNewBlock = (index = blocks.length, moveContentsDown = false) => {
-    const temp = blocks.map(block => ({...block}));
+    setBlocks(blocks => {
+      blocks.splice(index, 0, createBlock())
 
-    temp.splice(index, 0, getNewBlock())
+      const [colPos] = col;
 
-    const [colPos] = col;
+      if(moveContentsDown && index > 0) {
+        blocks[index].content += blocks[index-1].content.substring(colPos)
+        blocks[index-1].content = blocks[index-1].content.substring(0, colPos)
+      }
 
-    if(moveContentsDown && index > 0) {
-      temp[index].content += temp[index-1].content.substring(colPos)
-      temp[index-1].content = temp[index-1].content.substring(0, colPos)
-    }
+      return blocks
+    })
 
-    setBlocks(temp);
     setRow(index);
     setCol([0, 0]);
 }
 
   const onClick: MouseEventHandler<HTMLDivElement> = e => {
     if(!dragging) {
-      if(!blocks.length || blocks[blocks.length-1].content)
+      const last = blocks[blocks.length-1]
+      if(!last || last.content || last.isMetaBlock())
         addNewBlock();
       else
         setRow(row => row + 1);
@@ -86,8 +85,14 @@ const Editor: FC<Props> = ({ className, ...props }) => {
         break;
       case "Backspace":
         const [colPos, colLen] = col
+        const block = blocks[row];
         if(colPos == 0 && colLen == 0) {
-          deleteCurrentBlock();
+          if(block.isDefaultType()) 
+            deleteCurrentBlock();
+          else
+            modifyBlock(row, {
+              type: "p"
+            })
           e.preventDefault();
         }
         break;
@@ -103,55 +108,42 @@ const Editor: FC<Props> = ({ className, ...props }) => {
 
   useEffect(() => {
     window.addEventListener("mousedown", () => {
-      setBlocks(blocks => blocks.map(block => ({
-        ...block,
-        selected: false
-      })))
+      setSelected(-1)
     })
   }, [])
 
-  return <DragDropContext onDragStart={dragStart} onDragEnd={handleDrop}>
+  return <DragDropContext onDragStart={dragStart} onDragEnd={dragEnd}>
     <Droppable droppableId="list-container">
     {
-      (provided) => ( 
+      provided => ( 
         <div 
           onClick={onClick}
           onKeyDown={onKeyDown}
           className={`${styles.container}  ${className || ""}`}
           suppressContentEditableWarning={true}
           contentEditable={row >= 0}
-          { ...props }
-          { ...provided.droppableProps }
           ref={provided.innerRef}
-        > {
-          blocks.map((blockInfo, key) => {
-            return <Draggable key={blockInfo.id} draggableId={blockInfo.id} index={key}>
-              {
-                (provided) => (
-                  <div
-                    className="item-container"
-                    ref={provided.innerRef}
-                    {...provided.draggableProps} 
-                  >
-                    <Block
-                      blockInfo={blockInfo}
-                      modifyBlock={info => modifyBlock(key, info)} 
-                      focused={row==key}
-                      focus={() => setRow(key)}
-                      col={col}
-                      setCol={setCol}
-                      handleProps={provided.dragHandleProps}
-                      selected={selected == key}
-                      select={() => setSelected(key)}
-                    ></Block>
-                  </div>
-                )
-              }
-            </Draggable>
-          })
-        } 
+          { ...provided.droppableProps }
+          { ...props }
+        > 
+          {
+            blocks.map((block, key) => (
+              <BlockComponent
+                key={key}
+                index={key}
+                block={block}
+                modifyBlock={info => modifyBlock(key, info)} 
+                focused={row==key}
+                focus={() => setRow(key)}
+                col={col}
+                setCol={setCol}
+                selected={selected == key}
+                select={() => setSelected(key)}
+              ></BlockComponent>
+            ))
+          } 
 
-        {provided.placeholder} 
+          {provided.placeholder} 
         </div>
       )
     }
